Use iosif as only device CLI.
Pull screen dimensions and scale directly from device via MobileGestalt instead of requiring them to be entered into config.

Stop depending on mobiledevice and ios-deploy binaries and instead depend on new iosif tool.

Implement home and swipe.

Modularize the external binary used to interact with the phone itself so that different implementations of that can be made. ( so that after this go-ios and todevice can be supported for full cross platform ability )

Attempt at using screenshots when video is not received from the app. This doesn’t actually make the difference desired, which was to show system alerts. When system alerts occur the app still sends video… the alerts are just invisible :(

Move to updated version of json parser because the old version didn’t properly support single quoted keys and values nor did it work properly with capital case unquoted keys.

Revamped the device_tracker a bit to now require it’s own thread and channel. Therefore changed the setup to not use a goroutine for the http; so that everything will stay running with the main http “loop”.

Moved NewDevice function properly into the device.go file.

Simplified the way process shutdown is down via proc_generic a bit by the idea of a “procTracker”. This is actually still not quite right. If you unplug and replug a device the provider crashes. TODO