WDA updates. Show clean device type name
Renamed video app to “vidstream” insead of “vidtest”. This matches with the latest changes to the app itself.

Added settings in default.json for the video app name and bundle ids in case someone wants/needs to use customized ones.

Fetch the clean device type name now when starting up and pass it to CF for display.

Added an ‘-id’ option for wda mode, so that testing starting wda can be done for a single device when there are multiple devices connected.

Added a new ‘winsize’ command, so that the window size can easily be determined. This will be used in a new script I am writing to automatically create a proper devices block for config.json.

Improved cleanup to remove old tidevice processes.

Switched to using a “full” session creation for WDA instead of the shorter one that doesn’t specify a bundle id. This is because on some devices the short method doesn’t work at all. ( for example, it doesn’t work on XS Max )

Added all variety of error handling/retries for WDA functions. This *still* doesn’t resolve all the problems with WDA but it is at lest better than it was previously.

Added a long touch method and use it now for going to Screen Recording instead of firm press. Firm press works on some devices but the proper generic way for all devices is long press.