Use updated video app
The updated video app makes it possible to start video streaming without using the Control Center.
This new method is easier to trigger via WDA.
The updated code to start streaming is far more reliable and consistent in actually working.
It is still a bit glitchy and requires a sleep ( which should not be necessary imo... ), but that is due simply to how WDA works.