Improve keyboard handling.
Added monitoring of dasd to detect when application is changed.

Enabled mjpeg for WDA. It isn’t used yet but may be in the future as another available video mechanism. This depends / needs the latest version of tidevice.

Made keyMethod configurable but really right now I’ve only tested it with the new default configuration. Be warned if you fiddle with the setting. It may/will cause it to break.

Altered keys call to respond back, so that the frontend can wait for completion of the request.

On app changed a call is made to WDA to update the “tempApplication”. This is done so that keyboard entry can be faster and skip figuring out the application on every keystroke. As a result new updated WDA is required and the versionMarker for WDA has been updated.