Use new updated NNG based WDA; also various
Pass dev config forward into the device object for ease of referencing device specific configuration there.

Added multiple implementations of port proxying, so that different ones can be tested and/or different ones can be used if any of them are found to have a problem.

Listen to a different log message for app change, so that the specific app being brought to the foreground can be fetched and passed along.

Added new “manual” method for starting WDA, so that you can start WDA in Xcode and see crashes in it easily.

Changed to using a port range for the ports used by the provider instead of a hardcoded list. This will allow using many devices now. The port range is 8101-8200 by default.

Added a device specific config for the WDA port to use locally. This is meant to make it easier to integrate other things using WDA.

Added a ‘source’ command to dump the source of whatever is currently on screen for a device.

Changed to use NNG instead of HTTP for communicating with “WebDriverAgent”. The updated “WebDriverAgent” is very different though, and not really a “Web Driver Agent” any more. This change and the related changes within WDA itself fixes all manner of issues and makes all the actions a lot faster and more reliable.

This update needs updated video app and updated WebDriverAgent.