Avoid gestalt call if width/height set via config
Avoiding call to mobileGestalt if the info is set in the config.
The reason for this is because that call falls on some devices.
This change will cause width/height to be 0 in the displayed info though; will need to make a better change to not crash on devices where the call fails.