Use go-ios instead of iosif. Improved video streaming.
Disabled iosif from being pulled/built automatically as it is no longer the default.
Go-ios is now used primarily.

Now connecting to and outputting log stream from new ios video app. This allows for easier
debugging of what is going on with the video app.

Output error text from a failed generic process on startup.

Tweaks to video startup process to only have streaming active when a user is actively using the device.