Auto-respond to alerts ability. Shutdown tweaks. Various.
Remove output of tunnel "errors" as they are not really errors and are just log spam and confusing as they say they are errors in the output.

Added kill command to bridge; but did not yet implement it for iosif. As a result this commit is broken if you try to use it with iosif.

Auto-response ability for system alerts. Configurable.

Accept new shutdown command from CF; used to let CF shutdown the provider.

Kill vidstream extension on shutdown. This causes an error on the device, but it is better than leaving a running video stream.

Automatically dismiss alerts when starting up video stream, especially the leftover error from a previous shutdown.

Cleaned up various CLI commands and made them work again. Temporarily removed clickEl command; TODO add back.

Added stopChan passing for when WDA NNG channel should be stopped. This should make it possible to more cleanly stop WDA connection.