Add video app startup options
Added vidStartMethod per device, to allow Control Center to be used to start video streaming instead of via the app if desired.

The latest fixes to the video streaming app make activation via the app work on all devices, so this is unnecessary, but since it was broken I am going to leave this as an option in case it is needed again in the future.

Changed the schebang line for the python scripts so that it works on more machines regardless of how python was installed.

Fixed the WDA ElLongTouch cmd to work and match the current updated WDA.